Add jwt refresh

+++ src/Controller/AuthController.php
@@ -5,6 +5,7 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Annotation\Route;
 use App\Entity\User;
+use App\Entity\RefreshToken;
 use App\Controller\JWTauth;
 use Doctrine\Persistence\ManagerRegistry;
 use Symfony\Component\HttpFoundation\Response;
@@ -29,8 +30,11 @@ class AuthController extends AbstractController{
         }
 
         $passwordHash = password_hash($content['password'], PASSWORD_BCRYPT, ['cost' => 13]);
-        if(!isset($content['firstName']) || !isset($content['lastName'])){
-            return $this->json(['message' => 'User firstName and lastName cannot be empty'], 400);
+        if(!isset($content['firstName'])){
+            $content['firstName'] = "";
+        }
+        if(!isset($content['lastName'])){
+            $content['lastName'] = "";
         }
         
         $user = new User($content['username'], $content['email'], $passwordHash, $content['firstName'], $content['lastName']);
@@ -47,6 +51,17 @@ class AuthController extends AbstractController{
             return $this->json(['message' => 'You are already logged in']);
         }
         $content = $req->toArray();
+
+        $token = $doctrine->getRepository(RefreshToken::class)->findOneBy(['token' => $req->cookies->get('REFRESH')]);
+        if($token){
+            JWTauth::delete($doctrine, $req->cookies->get('REFRESH'));
+        }else{
+            $user = $doctrine->getRepository(User::class)->findOneBy(['username' => $content['username']]);
+            $token = $doctrine->getRepository(RefreshToken::class)->findOneBy(['usertoken' => $user]);
+            if($token){
+                JWTauth::delete($doctrine, $token->getToken());
+            }     
+        }
 
         $user = $doctrine->getRepository(User::class)->findOneBy(['username' => $content['username']]);
         $hash = $user->getPassword();
@@ -69,10 +84,25 @@ class AuthController extends AbstractController{
         }
     }
 
+    #[Route('/refresh', name: 'refresh', methods:['POST'])]
+    public function refresh(ManagerRegistry $doctrine, Request $req): Response{
+        if(is_null($req->cookies->get('REFRESH'))){
+            return $this->json(['message' => 'You\'re not logged in']);
+        }
+        $token = $doctrine->getRepository(RefreshToken::class)->findOneBy(['token' => $req->cookies->get('REFRESH')]);
+        if($token){
+            $user = $token->getUsertoken();
+            JWTauth::issueJWT($user->getUsername());
+            return $this->json(['message' => 'success']);
+        }else{
+            return $this->json(['message' => 'Invalid refresh token'], 403);
+        }
+    }
+
     #[Route('/reset', name: 'reset', methods:['PUT'])]
     public function reset(ManagerRegistry $doctrine, Request $req): Response{
         if($req->cookies->get('BEARER')){
-            return $this->json(['message' => 'You are already logged in'], 400);
+            return $this->json(['message' => 'You are already logged in'], 403);
         }
         $content = $req->toArray();
         $user = $doctrine->getRepository(User::class)->findOneBy(['email' => $content['email']]);
@@ -92,12 +122,10 @@ class AuthController extends AbstractController{
         if(!$req->cookies->get('BEARER')){
             return $this->json(['message' => 'You\'re not logged in'], 403);
         }
-        $return = '';
-        $test = JWTauth::verify($req->cookies->get('BEARER'));
-        foreach($test as $thing){
-            $return .= $thing;
+        $token = $doctrine->getRepository(RefreshToken::class)->findOneBy(['token' => $req->cookies->get('REFRESH')]);
+        if($token){
+            JWTauth::delete($doctrine, $req->cookies->get('REFRESH'));
         }
-        JWTauth::delete($doctrine, $req->cookies->get('REFRESH'));
-        return $this->json(['message' => $return], 200);
+        return $this->json(['message' => $token], 200);
     }
 }
